Add authorization for delete user endpoint

+++ DeleteUser/__init__.py
@@ -3,13 +3,20 @@ from shared import db
 import azure.functions as func
 import json
 from typing import Optional
+from shared.auth import auth
 
 
 def main(req: func.HttpRequest) -> func.HttpResponse:
     id = req.route_params.get('id')
     logging.info(f"Delete User (id: {id}) function processed a request.")
+    auth_data = auth(req)
 
-    return delete_user(id)
+    if(auth_data and auth_data["is_admin"]):
+        return delete_user(id)
+    else:
+        return func.HttpResponse(f"Unauthorized", status_code=403)
+
+
 
 def delete_user(id: int) -> func.HttpResponse:
     connection = db.get_connection()
